feat: Move Request server util method to RequestExt
