+++ examples/src/dynamic/server.rs
@@ -1,5 +1,9 @@
 use std::env;
-use tonic::{service::RoutesBuilder, transport::Server, Request, Response, Status};
+use tonic::{
+    service::RoutesBuilder,
+    transport::{RequestExt, Server},
+    Request, Response, Status,
+};
 
 use hello_world::greeter_server::{Greeter, GreeterServer};
 use hello_world::{HelloReply, HelloRequest};
