+++ examples/src/richer-error/server_vec.rs
@@ -1,4 +1,7 @@
-use tonic::{transport::Server, Code, Request, Response, Status};
+use tonic::{
+    transport::{RequestExt, Server},
+    Code, Request, Response, Status,
+};
 use tonic_types::{BadRequest, Help, LocalizedMessage, StatusExt};
 
 use hello_world::greeter_server::{Greeter, GreeterServer};
