+++ examples/src/streaming/server.rs
@@ -5,7 +5,10 @@ pub mod pb {
 use std::{error::Error, io::ErrorKind, net::ToSocketAddrs, pin::Pin, time::Duration};
 use tokio::sync::mpsc;
 use tokio_stream::{wrappers::ReceiverStream, Stream, StreamExt};
-use tonic::{transport::Server, Request, Response, Status, Streaming};
+use tonic::{
+    transport::{RequestExt, Server},
+    Request, Response, Status, Streaming,
+};
 
 use pb::{EchoRequest, EchoResponse};
