+++ examples/src/tls_client_auth/server.rs
@@ -3,7 +3,7 @@ pub mod pb {
 }
 
 use pb::{EchoRequest, EchoResponse};
-use tonic::transport::{Certificate, Identity, Server, ServerTlsConfig};
+use tonic::transport::{Certificate, Identity, RequestExt, Server, ServerTlsConfig};
 use tonic::{Request, Response, Status};
 
 type EchoResult<T> = Result<Response<T>, Status>;
